An error occured, when using histwin in Vim < 7.3.005
This was caused by the fact, that from 7.3.005 histwin uses the
undotree() function, which does not reliably work before 7.3.005
and so for older Vim versions histwin parses internally the :undolist
command. Unfortunally this would could an error for buffers, that
do not contain undo history (such as newly created buffers).

That should be fixed now and therefore, this is release 0.16